backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
anthropic==0.34.2
pypdf==4.3.1
python-docx==1.1.2
chardet==5.2.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
slowapi==0.1.9
python-json-logger==2.0.7
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-json-logger==2.0.7
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0